requirements.txt
fastapi==0.115.8
paramiko==3.5.1
requests==2.32.3
pyinstaller==6.12.0
uvicorn==0.34.0
websockets==15.0
pystray>=0.19.2
Pillow>=9.0.0
jinja2
pygame==2.6.1
SpeechRecognition==3.14.1
